[Company][Juan Carlos Salvador] feat: implemented getCompaniesByPagination API with JPA and change Parameter variable type from Long to Integer

+++ src/main/java/com/thoughtworks/springbootemployee/controller/CompanyController.java
@@ -33,7 +33,7 @@ public class CompanyController {
     }
 
     @GetMapping(params = {"pageIndex", "pageSize"})
-    public List<Company> getCompanyByPagination(@RequestParam Long pageIndex, @RequestParam Long pageSize) {
+    public List<Company> getCompanyByPagination(@RequestParam Integer pageIndex, @RequestParam Integer pageSize) {
         return companyService.getCompaniesByPagination(pageIndex, pageSize);
     }
